Review feedback: break out `example_chains` to crate
Accordingly, we need to move our benchmarks out to go with it. I've
taken this chance to throw out a few more focussing on different packet
parse types.
