+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
     "ingot",
+    "ingot-examples",
     "ingot-macros",
     "ingot-types",
 ]
