+++ ingot-examples/src/choices.rs
@@ -1,0 +1,34 @@
+//! Example uses of the [`choice`] macro.
+
+use ingot::{
+    choice,
+    ethernet::Ethertype,
+    icmp::{IcmpV4, IcmpV6, ValidIcmpV4, ValidIcmpV6},
+    ip::{IpProtocol, Ipv4, Ipv6, ValidIpv4, ValidIpv6},
+    tcp::{Tcp, ValidTcp},
+    udp::{Udp, ValidUdp},
+};
+
+/// An IPv4 or IPv6 header, determined by an input [`Ethertype`].
+#[choice(on = Ethertype)]
+pub enum L3 {
+    Ipv4 = Ethertype::IPV4,
+    Ipv6 = Ethertype::IPV6,
+}
+
+/// A TCP or UDP header, determined by an [`IpProtocol`] from an IPv4/v6
+/// packet.
+#[choice(on = IpProtocol)]
+pub enum L4 {
+    Tcp = IpProtocol::TCP,
+    Udp = IpProtocol::UDP,
+}
+
+/// An upper-layer protocol header: [`L4`], including ICMP(v6).
+#[choice(on = IpProtocol)]
+pub enum Ulp {
+    Tcp = IpProtocol::TCP,
+    Udp = IpProtocol::UDP,
+    IcmpV4 = IpProtocol::ICMP,
+    IcmpV6 = IpProtocol::ICMP_V6,
+}
