+++ ingot-examples/src/lib.rs
@@ -1,0 +1,10 @@
+//! Example uses of the [`Parse`] and [`choice`] macros.
+//!
+//! [`Parse`]: ingot::Parse
+//! [`choice`]: ingot::choice
+
+pub mod choices;
+pub mod packets;
+
+#[cfg(test)]
+mod tests;
