+++ ingot/Cargo.toml
@@ -1,7 +1,9 @@
 [package]
 name = "ingot"
 version = "0.1.0"
-edition = "2021"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
 
 [features]
 alloc = ["serde?/alloc"]
