+++ ingot/benches/modify.rs
@@ -1,13 +1,11 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use ingot::{
-    example_chain::{
-        GenericUlp, GeneveOverV6Tunnel, UdpParser, ValidGenericUlp,
-        ValidGeneveOverV6Tunnel, ValidUdpParser,
-    },
-    udp::{UdpMut, UdpRef, ValidUdp},
+    geneve::ValidGeneve,
+    ip::{Ipv4Flags, Ipv4Mut, Ipv4Ref, ValidIpv4},
+    types::HeaderParse,
+    udp::ValidUdp,
 };
-use ingot_types::HeaderParse;
-use std::{collections::LinkedList, hint::black_box};
+use std::hint::black_box;
 
 fn parse_udp(buf: &[u8]) -> ValidUdp<&[u8]> {
     ValidUdp::parse(buf).unwrap().0
@@ -15,81 +13,25 @@ fn parse_udp(buf: &[u8]) -> ValidUdp<&[u8]> {
 
 pub fn criterion_benchmark(c: &mut Criterion) {
     #[rustfmt::skip]
-    let pkt_body_v4: &mut[u8] = &mut [
-        // eth src : 0
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        // eth dst
-        0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-        // Ethertype
-        0x08, 0x00,
-        // IPv4 : 14
-        0x45, 0x00, 0x00, 28 + 8,
-        0x00, 0x00, 0x00, 0x00,
-        0xf0, 0x11, 0x00, 0x00,
-        192, 168, 0, 1,
-        192, 168, 0, 255,
-        // UDP : 34
+    let pkt_body_udp: &mut [u8] = &mut [
         0x00, 0x80, 0x17, 0xc1,
         0x00, 0x08, 0x00, 0x00,
-        // body : 42
-        0x00, 0x01, 0x02, 0x03,
-        0x04, 0x05, 0x06, 0x07,
     ];
+
     #[rustfmt::skip]
-    let pkt_body_v6: &mut [u8] = &mut [
-        // eth src
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        // eth dst
-        0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-        // Ethertype
-        0x86, 0xdd,
-        // v6
-        0x60, 0x00, 0x00, 0x00,
-        0x00, 0x10, 0x11, 0xf0,
-        // v6src
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-        // v6dst
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-        // UDP
-        0x00, 0x80, 0x17, 0xc1,
-        0x00, 0x08, 0x00, 0x00,
-        // body
-        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
+    let pkt_body_geneve_no_opt: &mut [u8] = &mut [
+        // ver + opt len
+        0x00,
+        // flags
+        0x00,
+        // proto
+        0x65, 0x58,
+        // vni + reserved
+        0x00, 0x04, 0xD2, 0x00,
     ];
+
     #[rustfmt::skip]
-    let opte_in_pkt = [
-        // ---OUTER ETH---
-        // dst
-        0xA8, 0x40, 0x25, 0x77, 0x77, 0x76,
-        // src
-        0xA8, 0x40, 0x25, 0x77, 0x77, 0x77,
-        // ethertype
-        0x86, 0xdd,
-
-        // ---OUTER v6---
-        // v6
-        0x60, 0x00, 0x00, 0x00,
-        0x00, 0x10, 0x11, 0xf0,
-        // v6src
-        0xFD, 0x00, 0x00, 0x00, 0x00, 0xF7, 0x01, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02,
-        // v6dst
-        0xFD, 0x00, 0x00, 0x00, 0x00, 0xF7, 0x01, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-
-        // ---OUTER UDP---
-        // source
-        0x1E, 0x61,
-        // dest
-        0x17, 0xC1,
-        // length
-        0x00, 0x14,
-        // csum
-        0x00, 0x00,
-
-        // ---GENEVE WITH OPT---
+    let pkt_body_geneve_opts: &mut [u8] = &mut [
         // ver + opt len
         0x01,
         // flags
@@ -105,76 +47,95 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         0x00,
         // rsvd + len
         0x00,
+    ];
 
-        // ---INNER ETH---
-        // dst (guest)
-        0xAA, 0x00, 0x04, 0x00, 0xFF, 0x10,
-        // src (gateway)
-        0xAA, 0x00, 0x04, 0x00, 0xFF, 0x01,
-        // ethertype (v4)
-        0x08, 0x00,
-
+    #[rustfmt::skip]
+    let pkt_body_ipv4_no_opt: &mut [u8] = &mut [
         // ---INNER v4---
         0x45, 0x00, 0x00, 28 + 8,
         0x00, 0x00, 0x00, 0x00,
         0xf0, 0x11, 0x00, 0x00,
         8, 8, 8, 8,
         192, 168, 0, 5,
-
-        // ---INNER UDP---
-        0x00, 0x80, 0x00, 53,
-        0x00, 0x08, 0x00, 0x00,
-
-        // ---INNER BODY---
-        0x00, 0x01, 0x02, 0x03,
-        0x04, 0x05, 0x06, 0x07,
     ];
-    let opte_out_pkt = &opte_in_pkt[opte_in_pkt.len() - 50..];
-    let mut actual_chain_v4 = LinkedList::new();
-    actual_chain_v4.push_front(pkt_body_v4[42..].to_vec());
-    actual_chain_v4.push_front(pkt_body_v4[34..42].to_vec());
-    actual_chain_v4.push_front(pkt_body_v4[14..34].to_vec());
-    actual_chain_v4.push_front(pkt_body_v4[0..14].to_vec());
+
+    #[rustfmt::skip]
+    let pkt_body_ipv4_opts: &mut [u8] = &mut [
+        // ---INNER v4---
+        0x49, 0x00, 0x00, 28 + 8,
+        0x00, 0x00, 0x00, 0x00,
+        0xf0, 0x11, 0x00, 0x00,
+        8, 8, 8, 8,
+        192, 168, 0, 5,
+
+        // as many bytes as we specified above the IHL.
+        0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15,
+    ];
 
     c.bench_function("parse-udp", |b| {
-        b.iter(|| black_box(parse_udp(black_box(&pkt_body_v4[34..42]))))
+        b.iter(|| black_box(parse_udp(black_box(pkt_body_udp))))
     });
-    c.bench_function("parse-stack-v4", |b| {
-        b.iter(|| UdpParser::parse(black_box(&pkt_body_v4[..])).unwrap())
-    });
-    c.bench_function("parse-and-decr-v4", |b| {
+
+    // Geneve and IP have varlen parts, IPV4's does not contain a subparse.
+    c.bench_function("parse-geneve-no-opt", |b| {
         b.iter(|| {
-            let (mut hdrs, ..) =
-                UdpParser::parse(black_box(&mut pkt_body_v4[..])).unwrap();
-            black_box(hdrs.l4.set_destination(hdrs.l4.destination() - 1));
+            ValidGeneve::parse(black_box(&pkt_body_geneve_no_opt[..])).unwrap()
         })
     });
-    c.bench_function("parse-stack-v6", |b| {
-        b.iter(|| UdpParser::parse(black_box(&pkt_body_v6[..])).unwrap())
-    });
-    c.bench_function("parse-valid-stack-v6", |b| {
-        b.iter(|| ValidUdpParser::parse(black_box(&pkt_body_v6[..])).unwrap())
-    });
-    c.bench_function("parse-read-v4", |b| {
+
+    c.bench_function("parse-geneve-opts", |b| {
         b.iter(|| {
-            UdpParser::parse_read(black_box(actual_chain_v4.iter())).unwrap()
+            ValidGeneve::parse(black_box(&pkt_body_geneve_opts[..])).unwrap()
         })
     });
-    c.bench_function("parse-stack-opte-in", |b| {
+
+    c.bench_function("parse-ipv4-no-opt", |b| {
         b.iter(|| {
-            GeneveOverV6Tunnel::parse(black_box(&opte_in_pkt[..])).unwrap()
+            ValidIpv4::parse(black_box(&pkt_body_ipv4_no_opt[..])).unwrap()
         })
     });
-    c.bench_function("parse-valid-opte-in", |b| {
+
+    c.bench_function("parse-ipv4-opts", |b| {
+        b.iter(|| ValidIpv4::parse(black_box(&pkt_body_ipv4_opts[..])).unwrap())
+    });
+
+    // Test speed needed to get/set/convert fields.
+    let (mut v4, ..) = ValidIpv4::parse(pkt_body_ipv4_no_opt).unwrap();
+
+    c.bench_function("get-int", |b| {
         b.iter(|| {
-            ValidGeneveOverV6Tunnel::parse(black_box(&opte_in_pkt[..])).unwrap()
+            black_box(v4.total_len());
         })
     });
-    c.bench_function("parse-stack-opte-out", |b| {
-        b.iter(|| GenericUlp::parse(black_box(&opte_out_pkt[..])).unwrap())
+
+    c.bench_function("set-int", |b| {
+        b.iter(|| {
+            v4.set_total_len(black_box(128));
+        })
     });
-    c.bench_function("parse-valid-opte-out", |b| {
-        b.iter(|| ValidGenericUlp::parse(black_box(&opte_out_pkt[..])).unwrap())
+
+    c.bench_function("get-int-unaligned", |b| {
+        b.iter(|| {
+            black_box(v4.version());
+        })
+    });
+
+    c.bench_function("set-int-unaligned", |b| {
+        b.iter(|| {
+            v4.set_version(black_box(4));
+        })
+    });
+
+    c.bench_function("get-networkrepr-flags", |b| {
+        b.iter(|| {
+            black_box(v4.flags());
+        })
+    });
+
+    c.bench_function("set-networkrepr-flags", |b| {
+        b.iter(|| {
+            v4.set_flags(black_box(Ipv4Flags::DONT_FRAGMENT));
+        })
     });
 }
